Modified test structure as per suggestion
Test-tag-hw-small: pr,hw,small dfusesparsefile
